% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_names.R
\name{add_parsed_names}
\alias{add_parsed_names}
\title{Add species and taxon names to a dataframe by
parsing the scientific names with gnparser}
\usage{
add_parsed_names(df, sci_name, ..., gnparser_path = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{sci_name}{Unquoted name of column in dataframe with
scientific name to be parsed}

\item{...}{Unquoted taxonomic names to add. May include:
species, taxon, taxon_with_rank, author, or year.}

\item{gnparser_path}{String; path to gnparser executable.
Either this must be provided, or gnparser must be on $PATH.}
}
\value{
Tibble
}
\description{
Requires gnparser (https://gitlab.com/gogna/gnparser/)
to be installed.
}
\examples{
names_df <- tibble::tibble(name = c(
  "Amaurorhinus bewichianus (Wollaston,1860) (s.str.)",
  "Crepidomanes minutum var. minutum"))
add_parsed_names(names_df, name, species, taxon)
add_parsed_names(names_df, name, species, author)
}
