% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{check_darwin_core_cols}
\alias{check_darwin_core_cols}
\title{Check data for presence and format of Darwin Core columns}
\usage{
check_darwin_core_cols(data)
}
\arguments{
\item{data}{Dataframe}
}
\value{
TRUE if passes test, or warning if not
}
\description{
Check data for presence and format of Darwin Core columns
}
\examples{
\dontrun{

data(filmy_taxonomy)
check_darwin_core_cols(filmy_taxonomy)

badly_formatted_data <- dplyr::select(filmy_taxonomy, -taxonRank)
check_darwin_core_cols(badly_formatted_data)

}
}
