% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{check_darwin_core_format}
\alias{check_darwin_core_format}
\title{Check that data used as taxonomic standard meets Darwin Core format}
\usage{
check_darwin_core_format(taxonomic_standard)
}
\arguments{
\item{taxonomic_standard}{Dataframe, should be meet Darwin Core format}
}
\value{
Tibble, or error if taxonomic_standard doesn't meet Darwin Core format
}
\description{
Check that data used as taxonomic standard meets Darwin Core format
}
\examples{
\dontrun{

data(filmy_taxonomy)
check_darwin_core_format(filmy_taxonomy)

badly_formatted_data <- dplyr::select(filmy_taxonomy, -taxonRank)
check_darwin_core_format(badly_formatted_data)

}
}
