% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{check_unique}
\alias{check_unique}
\title{Check for unique values in a dataframe}
\usage{
check_unique(df, ...)
}
\arguments{
\item{df}{Dataframe}

\item{...}{Bare name of column to check for unique values}
}
\value{
df If all values in column to check are unique,
error if not. If number of rows in df is zero, original
df will be returned.
}
\description{
Check for unique values in a dataframe
}
\examples{
\dontrun{
data <- tibble::tibble(a = c(1,2,3), b = c(2,2,4))
empty_data <- tibble::tibble()
check_unique(data, a)
check_unique(data, b)
check_unique(empty_data, a)
}

}
