% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filmy_taxonomy}
\alias{filmy_taxonomy}
\title{Taxonomy of filmy ferns (family Hymenophyllaceae)}
\format{
A data frame with 2729 rows and 31 variables.
}
\source{
\url{http://www.catalogueoflife.org/}
}
\usage{
filmy_taxonomy
}
\description{
A dataset containing taxonomic names and associated metadata for the
fern family Hymenophyllaceae. Downloaded from the
\href{http://www.catalogueoflife.org/}{Catalog of Life}, Version 1.5.
All columns formatted according to
\href{https://dwc.tdwg.org/terms/}{Darwin Core standard}. Only includes taxa
at the species or infraspecies level.
}
\keyword{datasets}
