% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{genus_name_only}
\alias{genus_name_only}
\title{Extract only the genus name from a longer name.}
\usage{
genus_name_only(taxon_name, sep = " ")
}
\arguments{
\item{taxon_name}{Taxon name, e.g. "Crepidomanes minutum var minutum".}

\item{sep}{Character separating parts of the name.}
}
\value{
The first two parts of the name separated by space.
}
\description{
It is assumed that the first part of the name is the genus.
No checking is done for this.
}
\examples{
\dontrun{
genus_name_only("Crepidomanes minutum var minutum")
}
}
