% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_taxonomy.R
\name{match_taxonomy}
\alias{match_taxonomy}
\title{Match species names to a reference database}
\usage{
match_taxonomy(query, taxonomic_standard, match_by = c("species",
  "taxon", "scientific_name"), max_dist = 0, simple = TRUE)
}
\arguments{
\item{query}{Character vector; taxonomic names to resolve.}

\item{taxonomic_standard}{Dataframe of standard names to match to.
Must follow \href{https://dwc.tdwg.org/terms/}{Darwin Core format}.}

\item{match_by}{One of "species", "taxon", or "scientific_name";
the type of standard name to use for matching.
Should match the type of names in \code{query}.}

\item{max_dist}{Integer; maximum distance to use during fuzzy matching.}

\item{simple}{Logical; should the results include only a limited set of
columns? If FALSE, all columns in \code{taxonomic_standard} will be returned.}
}
\value{
Dataframe. Includes the following columns:
\itemize{
\item \code{query}: the search query
\item \code{n_hits}: the number of matches to the query
\item \code{distance}: the number of characters differing between the query
and the match
\item \code{match_to}: the name in \code{taxonomic_standard} that matched the query
\item \code{match_by}: the type of species name that was used for matching
}

... and either all columns in \code{taxonomic_standard} or only a subset of
these if \code{simple} is TRUE.
}
\description{
Match a vector of species names to a user-provided taxonomic reference
database in the \href{https://dwc.tdwg.org/terms/}{Darwin Core format}.
}
\details{
For an example of a taxonomic reference database in Darwin Core format,
see \code{\link{filmy_taxonomy}}.

Fuzzy matching can be used by setting \code{max_dist} > 0. Fuzzy matching is
particularly useful for full scientific names (including author), as author
abbreviations and formatting vary widely.

It is assumed that all names in \code{query} are of the same type. Types of
species names include:
\itemize{
\item species: Species name not including any names below the
species level, e.g., "Iris pallida"
\item taxon: Species name including names below the species level (if any),
e.g., "Iris pallida cengialti". Note that the rank of the infraspecific
name (e.g., "var.") is not included.
\item scientific name: The full species name including infraspecific name (if any)
and author, e.g., "Iris pallida subsp. cengialti (Ambrosi ex A.Kern.) Foster".
}
}
\examples{
# Load reference taxonomy in Darwin Core format
data(filmy_taxonomy)

# This taxon matches many names at the species level because
# there are a bunch of varieties.
match_taxonomy("Hymenophyllum polyanthos", filmy_taxonomy, "species")

# Using the full species name with author can get us a
# more exact match.
match_taxonomy("Hymenophyllum polyanthos (Sw.) Sw.", filmy_taxonomy, "scientific_name")

# Fuzzy match helps when the query didn't abbreviate
# the author, but it is abbreviated in the reference.
match_taxonomy("Hymenophyllum polyanthos (Swartz) Swartz",
filmy_taxonomy, "scientific_name", max_dist = 8)

}
