% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_with_gnr.R
\name{match_with_gnr}
\alias{match_with_gnr}
\title{Match names using the Global Names Resolver}
\usage{
match_with_gnr(names, data_source_ids = 1, exclude_mult_matches = TRUE,
  gnparser_path = NULL)
}
\arguments{
\item{names}{Character vector of names to resolve}

\item{data_source_ids}{Numeric vector of data source IDs to use for
search. Default = 1, the Catalog of Life. For other data sources, see
\code{\link[taxize]{gnr_datasources}}}

\item{exclude_mult_matches}{Logical; should names resulting in multiple
matches be excluded as failures? If \code{TRUE} (the default), the output will have
a number of rows equal to length of \code{names}. If \code{FALSE}, the output may have
more rows than the length of \code{names}.}

\item{gnparser_path}{String; path to gnparser executable.
Either this must be provided, or gnparser must be on $PATH.}
}
\value{
tibble
}
\description{
Matches names using the \href{https://resolver.globalnames.org/}{Global Names Resolver}
API. This requires an internet connection. The GNR combines parsing of names
with fuzzy matching to match names to a target taxonomic database.
}
\details{
Attempts to match hybrid names usually don't work, so any names
that appear to by hybrids (e.g., with " x " in the name) are excluded

It is not recommended to use more than one data source at a time, as this may
complicate resolving synonymy.

Requires \href{https://gitlab.com/gogna/gnparser}{gnparser} to be installed.
}
\examples{
match_with_gnr(
 c(
  "Amaurorhinus bewichianus (Wollaston,1860) (s.str.)",
  "Crepidomanes minutum var. minutus",
  "Hymenophyllum ×polyanthos",
  "Acrostichum aureum",
  "fizz bazz")
)

match_with_gnr(
 c(
  "Amaurorhinus bewichianus (Wollaston,1860) (s.str.)",
  "Crepidomanes minutum var. minutus",
  "Hymenophyllum ×polyanthos",
  "Acrostichum aureum",
  "fizz bazz"),
  exclude_mult_matches = FALSE
)

}
