% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_names.R
\name{parse_names_batch}
\alias{parse_names_batch}
\title{Parse species names in batch}
\usage{
parse_names_batch(names, check = TRUE)
}
\arguments{
\item{names}{Character vector of species names.
May include author, variety, etc. All names must be
unique, with no NAs.}

\item{check}{Logical; should a check be made that the
results original name match the names of the input?}
}
\value{
Tibble
}
\description{
Runs much faster for a large number of names.
}
\details{
Requires gnparser to be installed and on $PATH
}
\examples{
parse_names_batch("Amaurorhinus bewichianus (Wollaston,1860) (s.str.)")
}
