% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_fern_names.R
\name{resolve_fern_names}
\alias{resolve_fern_names}
\title{Resolve names of ferns and lycophytes (pteridophytes)}
\usage{
resolve_fern_names(names, col_plants, exact_match_by = "scientific_name",
  resolve_to = c("species", "scientific_name"))
}
\arguments{
\item{names}{Vector of scientific names to resolve}

\item{col_plants}{Catalog of Life database subsetted to tracheophytes, in
Darwin Core format. This can be downloaded as part of a zip archive from
\href{http://www.catalogueoflife.org/DCA_Export/zip/archive-kingdom-plantae-phylum-tracheophyta-bl3.zip}{here}.
The unzipped file contains \code{taxa.txt}, which is a tab-separated file
including 1,168,025 observations of 31 variables, mostly taxonomic ranks
and names. \code{taxa.txt} should be read-in using \code{readr::read_tsv()} or similar,
and used as \code{col_plants}.}

\item{exact_match_by}{String, either "scientific_name" or "species". The first round of matching
will be done using exact matching to either scientific name (species plus author) or species (no author).
Default = "scientific_name".}

\item{resolve_to}{String, either "scientific_name" or "species". The type of
name that should be resolved. If "species", queries matching different varieties
within the same species will be collapsed to a single species.}
}
\value{
Tibble with the following columns
\itemize{
\item query: the query (\code{names}) after slight formatting (removing quotation marks)
\item query_original: the original query (\code{names})
\item non_pterido_genus: logical flag if name matched a non-pterdidophyte genus
\item taxonomicStatus: taxonomicStatus of the queried name
\item scientificName, genus, specificEpithet, species, infraspecificEpithet:
matched, resolved names (i.e., queries that were synonyms have been
replaced with the accepted name),
}
}
\description{
Excludes hybrids and any non-pteridophyte genera, and resolves names to
the \href{https://worldplants.webarchiv.kit.edu/ferns/}{World Ferns} list
contained within the \href{http://www.catalogueoflife.org/}{Catalog of Life}.
}
\details{
This takes ca. 10 minutes for 25,000 names.
}
\examples{
\dontrun{
# Specify location of Catalog of Life subsetted to tracheophytes
col_file <- "/Volumes/Transcend/Projects/pacific_ferns_biogeo/data_raw/archive-kingdom-plantae-phylum-tracheophyta-bl3/taxa.txt"

# Read in CoL. First get colnames, and use these to specify column type.
# taxonID, acceptedNameUsageID, parentNameUsageID need to be numeric.
col_plants_cols <- readr::read_tsv(col_file, n_max = 1) \%>\% colnames

col_spec <- dplyr::case_when(
  stringr::str_detect(col_plants_cols, "taxonID|acceptedNameUsageID|parentNameUsageID") ~ "n",
  TRUE ~ "c"
) \%>\% paste(collapse = "")

col_plants <- readr::read_tsv(
  col_file,
  col_types = col_spec)

names <- c("abacopteris insularis",
  "abacopteris philippinarum",
  "abacopteris rubra var. hirsuta",
  "acrostichum aureum",
  "cystopteris fragilis",
  "Cystopteris fragilis (L.) Bernh.",
  "Anemia stricta",
  "Athyrium macrocarpum",
  "Bulboschoenus fluviatilis",
  "Asplenium leptophyllum",
  "Asplenium leptophyllum baker",
  "Abrodictyum dentatum",
  "Abrodictyum dentatum Iwats.",
  "Polypodium vulgare",
  "Polypodium vulgare l. bak.",
  "Athyrium whazzat")

resolve_fern_names(names, col_plants, "scientific_name")
resolve_fern_names(names, col_plants, "species")

}

}
