% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_gnr_results.R
\name{resolve_gnr_results}
\alias{resolve_gnr_results}
\title{Resolve output of GNR matching}
\usage{
resolve_gnr_results(matched_names, taxonomic_standard)
}
\arguments{
\item{matched_names}{Dataframe of names matched to a taxonomic standard
(output of match_with_gnr).}

\item{taxonomic_standard}{Dataframe of standard names in Darwin Core format}
}
\value{
Tibble
}
\description{
After names have been parsed and fuzzily matched to names in a taxonomic
standard databse using \code{\link{match_with_gnr}}, this function can be used to resolve
synonyms. \strong{Very important:} the \strong{same} taxonomic standard should be used
for \code{\link{match_with_gnr}} and \code{\link{resolve_gnr_results}}.
}
\details{
Synonyms in matched names will be matched to the taxonomic standard.
If any are synonyms these will be replaced with the accepted name.
}
\examples{
# Load reference taxonomy in Darwin Core format
data(filmy_taxonomy)

gnr_match_results <-
match_with_gnr(c(
  "Cephalomanes oblongifolia",
  "Crepidomanes minutum var. minutus",
  "Hymenophyllum polyanthos Sw",
  "Abrodictyum extravagans (Copel.)",
  "adhga adf")
)

resolve_gnr_results(gnr_match_results, filmy_taxonomy)

}
