% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_gnr_results_by_rank.R
\name{resolve_gnr_results_by_rank}
\alias{resolve_gnr_results_by_rank}
\title{Resolve names matched with the Global Names Resolver to
either scientific name or species}
\usage{
resolve_gnr_results_by_rank(matched_names, taxonomic_standard,
  resolve_by_taxon = TRUE, resolve_by_species = TRUE,
  resolve_to = c("species", "scientific_name"))
}
\arguments{
\item{matched_names}{Tibble; output of matching names with Global Names Resolver}

\item{taxonomic_standard}{Tibble; taxonomic standard in Darwin Core format to
use for resolving names}

\item{resolve_by_taxon}{Logical; should names be resolved by matching to
taxon?}

\item{resolve_by_species}{Logical; should names be resolved by matching to
species?}

\item{resolve_to}{Taxonomic level to resolve names to; either "species"
or "scientific name"}
}
\description{
Resolve names matched with the Global Names Resolver to
either scientific name or species
}
