% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{sp_name_only}
\alias{sp_name_only}
\title{Extract only the species name from a longer name.}
\usage{
sp_name_only(taxon_name, sep = " ")
}
\arguments{
\item{taxon_name}{Taxon name, e.g. "Crepidomanes minutum var minutum".}

\item{sep}{Character separating parts of the name.}
}
\value{
The first two parts of the name separated by space.
}
\description{
It is assumed that the first two parts of the name are genus then
specific epithet. No checking is done for this.
}
\examples{
sp_name_only("Crepidomanes minutum var minutum")
}
