% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{toupper_first}
\alias{toupper_first}
\title{Convert the first letter of a string to uppercase}
\usage{
toupper_first(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector with first letter of
each element capitalized.
}
\description{
Convert the first letter of a string to uppercase
}
\examples{
\dontrun{
toupper_first(c("hi my name", "is"))
}
}
