% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_parse.R
\name{ts_parse}
\alias{ts_parse}
\title{Parse taxonomic names with (and for) taxon-tools}
\usage{
ts_parse(taxa)
}
\arguments{
\item{taxa}{Character vector; taxon names to be parsed by taxon-tools \code{parsenames}.
Missing values not allowed. Must all be unique.}
}
\value{
A dataframe including the following columns.
\itemize{
\item id: A unique ID number assigned to the input name
\item name: The input name
\item genus_hybrid_sign: Hybrid sign for genus
\item genus_name: Genus name
\item species_hybrid_sign: Hybrid sign for species
\item specific_epithet: Specific epithet (name)
\item infraspecific_rank: Infraspecific rank
\item infraspecific_epithet: Infraspecific epithet (name)
\item author: Name of taxon
}
}
\description{
Requires \href{https://github.com/camwebb/taxon-tools}{taxon-tools} to be installed.
}
\details{
Parses scientific names into their component parts (genus, species, variety, author, etc).
}
\examples{
\dontrun{
ts_parse("Foogenus x barspecies var. foosubsp (L.) F. Bar")
}

}
