% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_taxontools_installed.R
\name{ts_tt_installed}
\alias{ts_tt_installed}
\title{Test if taxon-tools is installed}
\usage{
ts_tt_installed()
}
\value{
\code{TRUE} if taxon-tools is installed, or \code{FALSE} if not.
}
\description{
A message will be issued if taxon-tools is not installed.
}
\examples{
ts_tt_installed()
}
