% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_write_names.R
\name{ts_write_names}
\alias{ts_write_names}
\title{Write out parsed names to a text file}
\usage{
ts_write_names(df, path)
}
\arguments{
\item{df}{Dataframe with parsed names}

\item{path}{Path to write dataframe

Writes out parsed names in a format that can be used by \href{https://github.com/camwebb/taxon-tools}{taxon-tools}
(each part of the scientific name is separated by the pipe symbol (|), with one name per line).}
}
\value{
Path to parsed names
}
\description{
Write out parsed names to a text file
}
\examples{
if (ts_tt_installed()) {
  parsed_names <- ts_parse_names(
    "Foogenus x barspecies var. foosubsp (L.) F. Bar")
  temp_file <- tempfile()
  ts_write_names(parsed_names, temp_file)
  readLines(temp_file)
  file.remove(temp_file)
}
}
