% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_with_gnr.R
\name{match_with_gnr}
\alias{match_with_gnr}
\title{Match names using the Global Name Resolver}
\usage{
match_with_gnr(names, data_source_ids = 1)
}
\arguments{
\item{names}{Character vector of names to resolve}

\item{data_source_ids}{Numeric vector of data source IDs to use for
search. Default = 1, the Catalog of Life. For other data sources, see
\code{\link[taxize]{gnr_datasources}}}
}
\value{
tibble
}
\description{
Attempts to match hybrid names usually don't work, so any names
that appear to by hybrids (e.g., with " x " in the name) are excluded
}
\details{
It is not recommended to use more than one data source at a time, and
default search values are set with this expectation.

Requires \href{https://gitlab.com/gogna/gnparser}{gnparser} to be installed
and on the user's PATH.
}
\examples{
match_with_gnr(c(
  "Amaurorhinus bewichianus (Wollaston,1860) (s.str.)",
  "Crepidomanes minutum var. minutus",
  "Hymenophyllum ×polyanthos ")
)

}
