% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_names.R
\name{parse_names_batch}
\alias{parse_names_batch}
\title{Parse species names in batch}
\usage{
parse_names_batch(names, gnparser_path = NULL)
}
\arguments{
\item{names}{Character vector of species names.
May include author, variety, etc. All names must be
unique, with no NAs.}

\item{gnparser_path}{String; path to gnparser executable.
Either this must be provided, or gnparser must be on $PATH.}
}
\value{
Tibble
}
\description{
Runs much faster for a large number of names.
}
\details{
Requires gnparser (https://gitlab.com/gogna/gnparser/)
to be installed.
}
\examples{
parse_names_batch("Amaurorhinus bewichianus (Wollaston,1860) (s.str.)")
parse_names_batch(c(
"Amaurorhinus bewichianus (Wollaston,1860) (s.str.)",
"bla",
"Amaurorhinus bewichianus (Wollaston,1860) (s.str.)")
)
}
