% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_hits.R
\name{resolve_hits}
\alias{resolve_hits}
\title{Sort through and resolve hits from taxonomic matching of a single name}
\usage{
resolve_hits(hits, col_to_resolve = c("scientific_name", "taxon",
  "species"), taxonomic_standard, accepted_singles_to_exclude = NULL,
  accepted_mults_to_exclude = NULL, synonym_singles_to_exclude = NULL,
  synonym_mults_to_exclude = NULL, mult_syn_selection = NULL,
  exclude = NULL, simple = TRUE)
}
\arguments{
\item{hits}{Dataframe; hits from taxonomic matching - output of
match_taxonomy().}

\item{col_to_resolve}{Name of column including taxonomic name to resolve.}

\item{taxonomic_standard}{Dataframe of standard names to match to.}

\item{accepted_singles_to_exclude}{Character vector; species that
should be excluded from single, accepted taxa.}

\item{accepted_mults_to_exclude}{Character vector; species that
should be excluded from multiple, accepted taxa.}

\item{synonym_singles_to_exclude}{Character vector; species that
should be excluded from single synonyms.}

\item{synonym_mults_to_exclude}{Character vector; species that
should be excluded from multiple synonyms.}

\item{mult_syn_selection}{Character vector of species names that should be
selected from those with multiple synonyms to use as the synonym
for that species.}

\item{exclude}{Character vector; taxa that should be exluded from the final
results.}

\item{simple}{Logical; should the results include only a limited set of
columns? If FALSE, all columns in \code{taxonomic_standard} will be returned.}
}
\value{
List including two items; \code{resolved_matches} is dataframe of resolved
names (including both accepted named and synonyms); \code{multiple_synonyms} are taxa
with multiple synonyms that could not be resolved.
}
\description{
Sort through and resolve hits from taxonomic matching of a single name
}
\examples{
# Load reference taxonomy in Darwin Core format
data(filmy_taxonomy)

taxonomic_standard <- filmy_taxonomy

# Single accepted name
tax_matching_results <- match_taxonomy(
  "Hymenophyllum polyanthos",
  filmy_taxonomy, "species")
resolve_hits(tax_matching_results, "species", filmy_taxonomy)

# Single synonym
tax_matching_results <- match_taxonomy(
  "Trichomanes crassum",
  filmy_taxonomy, "species")
resolve_hits(tax_matching_results, "species", filmy_taxonomy)

# Fuzzy match, results in both accepted name and multiple synonyms
tax_matching_results <- match_taxonomy(
  "Cephalomanes javanicam",
  filmy_taxonomy, "species", max_dist = 2)
resolve_hits(tax_matching_results, "species", filmy_taxonomy)
}
