% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_names_full.R
\name{resolve_names_full}
\alias{resolve_names_full}
\title{Resolve names based on scientific name, taxon, and species name}
\usage{
resolve_names_full(names_to_resolve, taxonomic_standard, exclude = NULL,
  match_order = c("sciname", "taxon", "species"),
  max_dist = c(scientific_name = 7, taxon = 2, species = 1),
  syn_select = NULL, check_key = TRUE)
}
\arguments{
\item{names_to_resolve}{Dataframe of taxonomic names to resolve.}

\item{taxonomic_standard}{Dataframe of standard names to match to. Here we use
the World Ferns database.}

\item{exclude}{Optional list of taxa vectors to exclude from matches. Names in each vector
must match names in \code{col_to_resolve} exactly.}

\item{match_order}{Vector designating order in which matching should proceed.
Names in vector must match column names in \code{names_to_resolve} used for matching.}

\item{max_dist}{Integer; maximum distance to use during fuzzy matching.}

\item{syn_select}{List of character vectors. In the case that multiple synonyms
were detected during matching, use the names specified. Names in each vector must match names in
\code{scientificName} column of \code{taxonomic_standard} exactly.}

\item{check_key}{Logical; should a check performed before joining resolved
and unresolved names that all name keys are unique? TRUE by default. If set
to FALSE, may cause repeat entries, so use with care!}
}
\value{
Tibble
}
\description{
Resolve names based on scientific name, taxon, and species name
}
\examples{
# Load reference taxonomy in Darwin Core format
data(filmy_taxonomy)

names_df <- tibble::tibble(
  sciname = c("Lycopodium kinabaluense Ching", "Gonocormus minutus (Blume) Bosch"),
  species = c("Lycopodium kinabaluense", "Gonocormus minutus")
)

resolve_names_full(
  names_to_resolve = names_df,
  taxonomic_standard = filmy_taxonomy,
  match_order = c("sciname", "species"),
  max_dist = c(
    sciname = 0,
    species = 0
  )
)

resolve_names_full(
  names_to_resolve = names_df,
  taxonomic_standard = filmy_taxonomy,
  match_order = c("sciname"),
  max_dist = c(
    sciname = 0
  )
)

}
