% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_names_full.R
\name{resolve_names_full}
\alias{resolve_names_full}
\title{Resolve names by recursive matching}
\usage{
resolve_names_full(names_to_resolve, taxonomic_standard,
  max_dist = c(scientific_name = 0, taxon = 0, species = 0))
}
\arguments{
\item{names_to_resolve}{Character vector; taxonomic names to resolve.}

\item{taxonomic_standard}{Dataframe of standard names to match to.}

\item{max_dist}{Named integer vector; maximum distance to use during
fuzzy matching by taxonomic rank.
Names of \code{max_dis} must include 'scientific_name', 'taxon', and 'taxon'.}
}
\value{
Tibble
}
\description{
The input (\code{names_to_resolve}) may include species names formatted
with or without author or infraspecific epithet. Names will be matched
recursively in order of scientific name (full name with author), taxon
name (name including infraspecific epithet but no author), then species
(genus plus specific epithet only).
}
\examples{
# Load reference taxonomy in Darwin Core format
data(filmy_taxonomy)

resolve_names_full(
  names_to_resolve = c(
    "Hymenophyllum polyanthos Sw.",
    "Crepidomanes minutum var. minutum",
    "Trichomanes crassum Copel.",
    "Hymenophyllum polyanthos Sw.",
    "whoops",
    "Nephrolepis brownii"),
  taxonomic_standard = filmy_taxonomy
)


}
