% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_resolve_names.R
\name{ts_resolve_names}
\alias{ts_resolve_names}
\title{Resolve synonyms in matched taxonomic names}
\usage{
ts_resolve_names(match_results, ref_taxonomy)
}
\arguments{
\item{match_results}{Dataframe; results of taxonomic name matching, most likely
done with \code{\link{ts_match_names}()}.}

\item{ref_taxonomy}{Dataframe; taxonomic data adhering to the \href{https://dwc.tdwg.org/terms/#taxon}{Darwin Core standard}
with the following columns:
\itemize{
\item \code{taxonID}: \href{https://dwc.tdwg.org/terms/#dwc:taxonID}{Unique identifier for each taxon}.
\item \code{acceptedNameUsageID}: \href{https://dwc.tdwg.org/terms/#dwc:acceptedNameUsageID}{An identifier for the name usage}
(documented meaning of the name according to a source) of the currently valid (zoological) or
accepted (botanical) taxon.
\item \code{taxonomicStatus}: \href{https://dwc.tdwg.org/terms/#dwc:taxonomicStatus}{The status of the use of the scientificName as a label for a taxon}.
Allowed values include \verb{accepted name}, \verb{ambiguous synonym}, \verb{provisionally accepted name}, \code{synonym}
\item \code{scientificName}: \href{https://dwc.tdwg.org/terms/#dwc:scientificName}{The full scientific name},
with authorship and date information if known.
}}
}
\value{
Dataframe; results of resolving synonyms in matched taxonomic names.
Includes the following columns:
\itemize{
\item \code{query}: Query name
\item \code{accepted_name}: Accepted name after resolving synonyms
\item \code{match_type}: Type of match (for a summary of match types, \href{https://github.com/camwebb/taxon-tools/blob/master/doc/matchnames.md#matching-rules-and-output-codes}{see taxon-tools manual})
\item \code{status}: Taxonomic status (same as \code{taxonomicStatus} in \code{ref_taxonomy})
\item \code{reference}: Matched reference name
}

Names that could not be matched or resolve to multiple, different synonyms have \code{NA} for \code{accepted_name}.
}
\description{
After matching query names to a reference taxonomy, some may match synonyms.
This function resolves synonyms to their accepted names.
}
\details{
\code{ref_taxonomy} must be taxonomic data adhering to the \href{https://dwc.tdwg.org/terms/#taxon}{Darwin Core standard}.
Darwin Core includes many terms, but only four (\code{taxonID}, \code{acceptedNameUsageID},
\code{taxonomicStatus}, and \code{scientificName}) are required for this function.

The scientific names in \code{ref_taxonomy} must be the same used as reference
with \code{\link{ts_match_names}()}.
}
\examples{
# Load reference taxonomy in Darwin Core format
data(filmy_taxonomy)

# Query a misspelled name
match_results <- ts_match_names(
  query = "Gonocormus minutum",
  reference = unique(filmy_taxonomy$scientificName),
  simple = TRUE)

# Resolve the synonyms
ts_resolve_names(match_results, filmy_taxonomy)
}
