% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_resolve_names.R
\name{ts_resolve_names}
\alias{ts_resolve_names}
\title{Resolve synonyms in taxonomic names}
\usage{
ts_resolve_names(
  query,
  ref_taxonomy,
  max_dist = 10,
  match_no_auth = FALSE,
  match_canon = FALSE,
  collapse_infra = FALSE,
  collapse_infra_exclude = NULL,
  docker = getOption("ts_docker", default = FALSE),
  tbl_out = getOption("ts_tbl_out", default = FALSE)
)
}
\arguments{
\item{query}{Character vector or dataframe; taxonomic names to be resolved.
If a character vector, missing values not allowed and all values must be
unique. If a dataframe, should be taxonomic names matched with
\code{\link{ts_match_names}()}.}

\item{ref_taxonomy}{Dataframe; reference taxonomic data adhering to the
\href{https://dwc.tdwg.org/terms/#taxon}{Darwin Core standard} with the
following columns:
\itemize{
\item \code{taxonID}: \href{https://dwc.tdwg.org/terms/#dwc:taxonID}{Unique identifier for each taxon}.
\item \code{acceptedNameUsageID}: If the taxon is a synonym, the \href{https://dwc.tdwg.org/terms/#dwc:acceptedNameUsageID}{unique identifier for the accepted name}
\item \code{taxonomicStatus}: \href{https://dwc.tdwg.org/terms/#dwc:taxonomicStatus}{The status of the use of the \code{scientificName} as a label for the taxon}.
\item \code{scientificName}: \href{https://dwc.tdwg.org/terms/#dwc:scientificName}{The full scientific name of the taxon},
with authorship and date information if known.
}}

\item{max_dist}{Max Levenshtein distance to allow during fuzzy matching
(total insertions, deletions and substitutions). Default: 10.}

\item{match_no_auth}{Logical; If no author is given in the query and the name (without author)
occurs only once in the reference, accept the name in the reference as a match.
Default: to not allow such a match (\code{FALSE}).}

\item{match_canon}{Logical; Allow a "canonical name" match if only the genus, species epithet,
and infraspecific epithet (if present) match exactly. Default: to not allow such a match (\code{FALSE}).}

\item{collapse_infra}{Logical; if the specific epithet and infraspecific epithet
are the same, drop the infraspecific rank and epithet from the query. For more
information, see \code{\link{ts_match_names}()}.}

\item{collapse_infra_exclude}{Character vector; taxonomic names to exclude
collapsing with \code{collapse_infra}. Any names used must match those in \code{query}
exactly, or they won't be excluded.}

\item{docker}{Logical; if TRUE, docker will be used to run taxon-tools
(so that taxon-tools need not be installed).}

\item{tbl_out}{Logical vector of length 1; should a tibble be returned?
If \code{FALSE} (default), output will be a data.frame. This argument can
be controlled via the option \code{ts_tbl_out}; see Examples.}
}
\value{
Dataframe; results of resolving synonyms in matched taxonomic names.
Includes the following columns:
\itemize{
\item \code{query}: Query name
\item \code{resolved_name}: Accepted name after resolving synonyms
\item \code{matched_name}: Name matched to query
\item \code{resolved_status}: Taxonomic status of the resolved name (same as \code{taxonomicStatus} in \code{ref_taxonomy})
\item \code{matched_status}: Taxonomic status of the matched name (same as \code{taxonomicStatus} in \code{ref_taxonomy})
\item \code{match_type}: Type of match (for a summary of match types, \href{https://github.com/camwebb/taxon-tools/blob/master/doc/matchnames.md#matching-rules-and-output-codes}{see taxon-tools manual})
}

Names that could not be matched or resolve to multiple, different synonyms have \code{NA} for \code{resolved_name}.
}
\description{
After matching taxonomic names to a reference, some may match synonyms. This
function resolves synonyms to their accepted names.
}
\details{
\code{query} can take as input either a character vector of taxonomic names, or
the output of \code{\link{ts_match_names}()}. If the former, it will run
\code{\link{ts_match_names}()} to match the query to \code{ref_taxonomy}, then
resolve synonyms. If the latter, the scientific names in \code{ref_taxonomy}
should be the same used as reference with \code{\link{ts_match_names}()}
(this is not checked).

\code{ref_taxonomy} must be taxonomic data adhering to the \href{https://dwc.tdwg.org/terms/#taxon}{Darwin Core standard}.
Darwin Core includes many terms, but only four (\code{taxonID},
\code{acceptedNameUsageID}, \code{taxonomicStatus}, and \code{scientificName}) are required
for this function.
}
\examples{
if (ts_tt_installed()) {
  # Load reference taxonomy in Darwin Core format
  data(filmy_taxonomy)

  ts_resolve_names("Gonocormus minutum", filmy_taxonomy)
  # If you always want tibble output without specifying `tbl_out = TRUE`
  # every time, set the option:
  options(ts_tbl_out = TRUE)
  ts_resolve_names("Gonocormus minutum", filmy_taxonomy)
}

}
