% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_tt_installed.R
\name{ts_tt_installed}
\alias{ts_tt_installed}
\title{Test if \href{https://github.com/camwebb/taxon-tools}{taxon-tools} is installed}
\usage{
ts_tt_installed()
}
\value{
\code{TRUE} if \href{https://github.com/camwebb/taxon-tools}{taxon-tools} is
installed, or \code{FALSE} if not.
}
\description{
Test if \href{https://github.com/camwebb/taxon-tools}{taxon-tools} is installed
}
\examples{
ts_tt_installed()
}
